refactor: 커스텀 훅 및 함수 분리를 통한 코드 개선 (#36)
