+++ src/app/page.tsx
@@ -1,5 +1,4 @@
 import Image from "next/image";
-// import { auth } from "@/auth"; // auth 추가
 import Header from "@/components/organism/layout/Header";
 import NavBar from "@/components/organism/layout/NavBar";
 import MainBanner from "@/app/(commonAccess)/_main/MainBanner";
@@ -7,56 +6,14 @@ import MainArticle from "@/components/organism/layout/MainArticle";
 import styles from "@/styles/main/main.module.scss";
 import BoardTitle from "@/components/organism/main/BoardTitle";
 import SimpleBoardObject from "@/components/organism/main/SimpleBoardObject";
-import CategoryText from "@/components/atoms/Text/CategoryText";
 import Footer from "@/components/organism/layout/Footer";
 import { auth } from "@/auth";
-
-async function getMainStatistic() {
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_REACT_APP_API_URL}/auction-service/api/v1/non-authorization/auction/statistic`,
-    {
-      next: { revalidate: 86400 },
-    }
-  );
-  if (!res.ok) {
-    throw new Error("Network Error");
-  }
-  const data = await res.json();
-  return data;
-}
-
-async function getMainHotActionList() {
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_REACT_APP_API_URL}/auction-service/api/v1/non-authorization/auction/hot-auction`
-  );
-  if (!res.ok) {
-    throw new Error("Network Error");
-  }
-  const data = await res.json();
-  return data;
-}
-
-async function getMainHighBiddingList() {
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_REACT_APP_API_URL}/auction-service/api/v1/non-authorization/auction/high-bidding-statistics`
-  );
-  if (!res.ok) {
-    throw new Error("Network Error");
-  }
-  const data = await res.json();
-  return data;
-}
-
-async function getMainSameCategoryList() {
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_REACT_APP_API_URL}/auction-service/api/v1/non-authorization/auction/category-hot-auction`
-  );
-  if (!res.ok) {
-    throw new Error("Network Error");
-  }
-  const data = await res.json();
-  return data;
-}
+import {
+  getMainStatistic,
+  getMainHotActionList,
+  getMainHighBiddingList,
+  getMainSameCategoryList,
+} from "@/utils/main/handleMainListData";
 
 export default async function Home() {
   // const session = await auth(); // session 호출 추가
@@ -81,7 +38,6 @@ export default async function Home() {
         title="HOT 경매글 📢"
         detail="지금 가장 인기있는 경매글을 찾아보세요!"
       />
-
       <div className="flex flex-col justify-center w-full h-200">
         <ul className="flex overflow-x-auto">
           {data2.map((item: any) => (
